test: Fix data deletion e2e tests
The "Delete MetaMetrics Data" e2e tests were recently broken due to a
change in CI configuration. The code-under-test was written to always
use the environment variable present for the data deletion source ID
and endpoint, but the e2e tests wrongly assumed that it would never be
present.

The service has been updated to use the fallback values for e2e test
builds, even if the environment variable is present.