Make SwapsController support multichain
Currently, SwapsController relies on the global network: it uses the
global provider from the NetworkController and feeds it into an Ethers
contract object to get the allowance for a user's token, and it uses the
chain ID of the global network (again, coming from NetworkController) to
retrieve parameters that will ultimately be used to control when and how
swaps are requested and how to transform the resulting data.

As part of the multichain initiative, we don't want to rely on the
global network anymore. So this commit updates anywhere this is
happening so that consumers must now pass a network client ID instead.
The list of changes is:

- Update the signature of the SwapsController constructor:
  - Remove `provider` as an option
  - Update type for `gasLayer1GasFee` option to be a function that takes
    `networkClientId` as a parameter instead of `chainId` (note: this
    function is a reference to a method in GasFeeController which
    already supports these arguments)
  - Update type for `fetchTradesInfo` option to be a function that takes
    `networkClientId` in `fetchMetadata` instead of `chainId`
- Update the `FetchTradesInfoParams` type to replace `chainId` with
  `networkClientId` (note: this also affects `SwapsControllerState`)
- Update `fetchAndSetQuotes` to take a `networkClientId` instead of
  `chainId` within the `fetchParamsMetaData` (second argument)
- Update `fetchAndSetQuotes` to no longer rely on a cached provider
  object (which is the provider proxy from NetworkController) but
  instead to always use the (non-proxy) provider from NetworkController
  that corresponds to the given network client ID
- Update signature for `getTopQuoteWithCalculatedSavings`: instead of
  taking a single quotes object, it now takes an options bag with keys
  `quotes` and `networkClientId`
- Add new selector `getSelectedNetwork` which can be used to get
  information about the global network configuration + network client ID
  in a Redux action without honoring the old "provider config" interface