fix: Make QR scanner more strict
The QR scanner is now more strict about the contents it allows to be
scanned. If the scanned QR code deviates at all from the supported
formats, it will retur "unknown" as the result (as it always has for
completely unrecognized QR codes).

Previously we would accept QR codes with a recognized prefix even if
the complete contents did not match our expectations, which has
resulted in unexpected behavior.