feat(SwapsController): Remove reliance on global network (#28275)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

Currently, SwapsController relies on the global network: it uses the
global provider from the NetworkController and feeds it into an Ethers
contract object to get the allowance for a user's token, and it uses the
chain ID of the global network (again, coming from NetworkController) to
retrieve parameters that will ultimately be used to control when and how
swaps are requested and how to transform the resulting data.

This does not work long-term, as we want to remove the concept of the
global network from the extension. So fundamentally, we need to have the
parts of the API which rely on the global provider to take a network
client ID, which will be used to fetch a provider and chain ID.

One important note about SwapsController is that there is a kind of
entry point to the API via `fetchAndSetQuotes`. This establishes the
network that a polling loop will use, and it starts that loop. The
controller also operates on one network at a time: it does not
distinguish between one network or another when capturing the quote data
in state, so when calling `fetchAndSetQuotes` for a different network,
if the polling loop was already started for a different network, it will
be reset, and all quotes captured for that network will be overwritten.
We plan on keeping this behavior.

Given these constraints, here is the list of changes:

- Update the `FetchTradesInfoParams` type to replace `chainId` with
`networkClientId`.
- Update `fetchAndSetQuotes` to take a `networkClientId` instead of
`chainId` within the `fetchParamsMetaData` (second argument). When
called for the first time, this ID will be resolved to a network client.
- Update `fetchAndSetQuotes` to no longer rely on the global provider
proxy from NetworkController to get the allowance for a token, but use
the network client obtained in the previous step.
- Update signature for `getTopQuoteWithCalculatedSavings`: instead of
taking a single quotes object, it now takes an options bag with keys
`quotes` and `networkClientId`. Additionally, instead of relying on the
global network to get the chain ID, it will resolve the network client
ID to a network client and get the chain ID off of that.
- Add new selector `getSelectedNetwork` which can be used to get
information about the global network configuration + network client ID
in a Redux action in one go (`getCurrentNetwork` was not sufficient to
do this).

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/28275?quickstart=1)

## **Related issues**

Progresses #23566.

## **Manual testing steps**

1. Switch to Ethereum Mainnet if you haven't already.
2. Click on the "Swap" button.
3. Open Dev Tools and go to the Network tab.
4. Fill in a monetary value for ETH.
5. Choose a destination token (it doesn't matter).
6. Wait for quotes to appear.
7. Observe that the text above the quotes says "Confirmed by X sources.
Verify on Etherscan."
8. In the Dev Tools pane, search for "swap.api". Find any request that
starts with `https://swap.api.cx.metamask.io/networks/`. Observe that
the chain ID in the URL matches Ethereum Mainnet.
9. Now search for "gas.api". Find any request that starts with
`https://gas.api.cx.metamask.io/networks/`. Observe that the chain ID in
the URL matches Ethereum Mainnet.
10. Cancel the swap.
11. Switch to Linea Mainnet.
12. Fill in a monetary value for ETH.
13. Choose a destination token.
14. Wait for quotes to appear.
15. Observe that the text above the quotes says "Confirmed by X sources.
Verify on LineaScan."
16. In the Dev Tools pane, search for "swap.api". Find any request that
starts with `https://swap.api.cx.metamask.io/networks/`. Observe that
the chain ID in the URL matches Linea Mainnet.
17. Now search for "gas.api". Find any request that starts with
`https://gas.api.cx.metamask.io/networks/`. Observe that the chain ID in
the URL matches Linea Mainnet.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

No screenshots/recordings necessary. Everything should work exactly the
same as it does today.

## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: MetaMask Bot <metamaskbot@users.noreply.github.com>