refactor: Allow getting manifest flags without setting
The pieces of `set-manifest-flags.ts` related to _getting_ flags have
been moved to the new `get-manifest-flags.ts` module. This module will
be used in a later PR by a script run during a CI workflow.

The migrated steps were also made asynchronous so that the asynchronous
steps could be run in parallel rather than blocking the process.

Documentation has been added to the `ManifestFlags` type as well, in
preparation for adding a new property in the next PR.