chore: Update `@metamask/gas-fee-controller` and peer deps
Update the `@metamask/gas-fee-controller` package to v21 to satisfy the
peer dependency on `@metamask/network-controller@^21.0.0`, and update
the `@metamask/user-operation-controller` to satisfy its peer
dependency upon `@metamask/gas-fee-controller`.

Note that an older version of `@metamask/gas-fee-controller` (v18)
remains in the dependency tree, but only because it's imported by
`@metamask/smart-transaction-controller` for type reasons. It has no
runtime impact on the application, so the associated peer dependency
warnings from this older release can be ignored. This will be
eliminated soon, in an upcoming PR.

The updated `@metamask/user-operation-controller` still does not have
its peer dependencies satisfied, but the problems are pre-existing. The
`@metamask/keyring-controller` and `@metamask/transaction-controller`
packages are head of where this package expects them to be. This is not
made worse by this PR though, and will be addressed in a future PR.