test: Fix `getEventPayloads` e2e test helper
The test helper `getEventPayloads` is meant to wait until all mock
endpoints provided have resolved (i.e. until none are pending).
However, it currently only waits until the _last_ mock endpoint in the
array has been fulfilled. The pending status of the others is ignored.
This results in race conditions, because the status of non-last mock
endpoints is indeterminate when this returns.

The `hasRequest` parameter has been renamed to `waitWhilePending` to
make the expected behavior more clear, and JSDocs have been added for
each parameter.