chore: Update CODEOWNERS to remove owner from some files
The CODEOWNERS file has been updated to remove owners from key files
that are frequently updated by multiple teams. Certain files that are
validated as part of the release process have also been unset, as there
are no risks with teams updating them in PRs.

Lastly, ownership of the CODEOWNERS file itself was moved to the
extension security team (it was accidentally set to privacy reviewers).