fix: Fix create metric fragment
The function `createEventFragment` of the `MetaMetricsController` was
broken recently in the migration to BaseControllerV2 (#28113). We ended
up trying to mutate a piece of Immer state, resulting in an error.

The affected line was updated to use an empty base object as the merge
target, so that we're no longer attempting to mutate a frozen object.

Fixes ##28599