refactor: Robustify migration 135 (#29168)
Robustifies migration 135 by using more of our own types throughout the
migration file. The result of reviewing migration 135 and ensuring that
it's functioning correctly. The only behavioral difference is that
`getChainIdForNetworkClientId()` will capture an exception if no chain
id can be found, as opposed to this occurring outside of that function.
The relevant test cases have been updated to ensure that we capture this
exception.