ci: Improve accuracy of `wait-for-circleci-workflow-status`
The GitHub Actions workflow `wait-for-circleci-workflow-status` has
been updated to ensure that it waits for the correct workflow.
Previously it always chose the most recent workflow for the given
branch, but it may have chosen a workflow corresponding to the wrong
commit. It has been updated to find one matching the same commit that
triggered the GitHub Actions workflow.