fix: Spending cap flicker as decimals are determined (#29206)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

Sometimes, when approving a token, the spending cap displayed in the
transaction simulation component flickers for a split second. This can
also seen when editing the spending cap on the same screen.

In the approve screen, `useAssetDetails` returns `decimals` at first as
`undefined` while it's determined asynchronously. Since
`useApproveTokenSimulation` takes `decimals` as an argument, a default
of `'0'` was set to quiet the type-checker. This default is what
provokes the UI flicker.

In the example video below, the token has 4 decimals and the spending
cap is 70000 / 10 ** 4 = 7. But while decimals is still `undefined`,
`'0'` is used in `useApproveTokenSimulation` to determine the spending
cap (dividing the value by 10 to the number of decimals). This amounts
to `70000` instead of `7` for a split second, before decimals `'4'` is
returned by `useAssetDetails`.

The fix for this bug is to let the loading spinner linger a few
miliseconds longer while decimals is still `undefined`.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/29206?quickstart=1)

## **Related issues**

Fixes:

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->



https://github.com/user-attachments/assets/a09f449a-78ea-4083-b47b-2f329126d4b6



### **After**

<!-- [screenshots/recordings] -->



https://github.com/user-attachments/assets/68142912-ae20-47f6-8dd4-2b9184b57bbf



## **Pre-merge author checklist**

- [ ] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.