chore: Remove broken coverage report link
The unit test coverage report was previously available in the
`metamaskbot` comment, but this link has been broken since we migrated
unit tests from CircleCI to GitHub Actions in #25570

The broken link has been removed for now. It can be restored in the
future, after migrating the `metamaskbot` comment to GitHub actions as
well.