Remove old properties from state
There are several kinds of errors in Sentry which indicate that there
are properties in controller state we are attempting to persist that do
not have corresponding metadata. This indicates that these properties
were removed at some point from the controller's state but no migration
was added that removed them from the persisted wallet state. In many
cases, at the time of removal, such a migration was not needed because
the controller in question inherited from BaseController v1. However, we
have made a targeted effort over the past few years to migrate all
controllers to BaseController v2, and so it matters now that every
property have corresponding metadata or else are removed from state. We
don't want these errors to show up in Sentry because they create noise,
so this commit removes these properties from state.