Upgrade queued-request-controller to v9, selected-network-controller to v21
These upgrades revert changes made in a previous major version which
would have removed the ability to disable per-dapp selected network.
This is not needed here on Extension, but is needed on Mobile. So this
commit doesn't have any material effect here but is merely to ensure
that extension is using the latest versions of these controllers.