fix: Fix bundle size diffs
Fix bundle size diffs that are included in `metamaskbot` comments and
tracked historically in https://github.com/MetaMask/extension_bundlesize_stats

The diffs were incorrect because the "merge base" commit was wrong. The
merge base commit is the commit we're comparing the branch against. We
were calculating this using `git merge-base origin/main HEAD`, which
locally does give the correct answer. But in GitHub Actions, HEAD was
not set to the tip of the branch. Instead it was a new comment
generated during the checkout step, representing the target branch
merged into the current branch.

This resulted in the "merge base" always being the tip of `main` (at
the time of this workflow run). This in turn resulted in the "bundle
diff" including changes that were already made on `main`

Fixes #29858.