fix: Lint `en_GB` locale
The `verify-locales` linter script now lints `en_GB`. Previously it was
ignored, so it had accumulated errors. All errors in this locale file
have been corrected as well.

This locale file is not actually used in the app, it's just present to
enable us to set a different UK-specific store description.

Closes #29965