ci: Take DOM snapshots of all windows on failure
Our e2e test tooling takes a DOM snapshot on test failure to help with
debugging. Unfortunately this is not overly useful right now because
the driver is always set to the last window handle when this snapshot
is taken, which is typically the offscreen document (at least on
Chrome). Even on Firefox where there is no offscreen document, it may
end up taking a snapshot of the wrong page when there are multiple
open.

The DOM snapshot step has been updated to take snapshots of all open
pages, rather than just the "current" one.