chore: Temporarily store MMI policy validation/update
Recently we removed all CI steps related to MMI, in preparation for
removing MMI. However, the lack of policy validation for the MMI
LavaMoat policy has led to churn/conflicts from local updates to the
policy.

The CI config has been updated to restore MMI LavaMoat policy
validation/updates temporarily. We expect we'll be able to delete this
again very soon, but it is easiest to keep this until we're ready to
delete the MMI build type.