fix: Ensure has_marketing_consent user trait are correctly set and sent after participateInMetaMetrics is set to true (#29871)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

It was observed in production metrics that the `has_marketing_consent`
user trait was not set (either `null` or `undefined`) for 40-50% of
users. Meanwhile, another ~40% of users had it set to false, and the
remainder had it set to true.

The problem was in the metametrics controller. The controller sets
`state.previousUserTraits` in the `_buildUserTraitsObject` function,
which gets called when there are state updates. That function also uses
`state.previousUserTraits` to determine which values need to be
returned, by comparing it to the new `currentTraits`.

The problem is that if a trait is set while `participateInMetaMetrics`
is false, then the trait will not be sent in a request to Segment, BUT
the trait will be saved in `state.previousUserTraits`. Later, if
`participateInMetaMetrics` is set to true, `_buildUserTraitsObject` will
not return a trait currently matches `state.previousUserTraits`

The `has_marketing_consent` trait is set to true when the user checks
the associated checkbox on the metametrics screen of the onboarding
flow, but then if the user clicks confirm/yes on that screen to opt-in
to metametrics, the aforementioned problem is hit.

To solve this, we just ensure that `state.previousUserTraits` is not
saved if `participateInMetaMetrics` is false.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/29871?quickstart=1)

## **Related issues**

Fixes: https://github.com/MetaMask/MetaMask-planning/issues/3932

## **Manual testing steps**

1. Install a build from this branch
2. Open the background console and the network tab
3. Start onboarding. On the metametrics screen, click the checkbox and
then "I agree"
4. The network tab should now include a request to segment with user
traits, where `has_marketing_consent` is set to true

--

1. Follow the above steps on the v12.9.3 build (step 4 will fail)
2. Update the version of that install to the build from this branch
3. Log in.
4. The network tab should now include a request to segment with user
traits, where `has_marketing_consent` is set to true

--

1. Install a build from this branch
2. Open the background console and the network tab
3. Start onboarding. On the metametrics screen, DO NOT click the
checkbox and then click "No thanks"
4. The network tab should not include a request to segment
5. Complete onboarding and go to settings and toggle "Participate in
MetaMetrics" to true
6. The network tab tab should now include a request to segment with user
traits, where `has_marketing_consent` is set to false. If the marketing
consent toggle is turned on, there should then be a network request
where the user trait is true

> [!IMPORTANT]
> This requires manual test from QA.

- Create a new user and click both "I agree" and "we'll use this data...
" when user is on "help us improve Metamask" page
- There should be 1 request to segment with user traits, where both
has_marketing_consent and participate_in_metametrics set to true.
- Proceed on onboarding
- Head to privacy page in settings
- Opt out "participate in metametrics"
- There's no more request to segment after

## **Screenshots/Recordings**



https://github.com/user-attachments/assets/0d335973-eb52-4f61-9709-2999efde4021



## **Pre-merge author checklist**

- [ ] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: dddddanica <zhaodanica@gmail.com>