feat(action): Improve bug report creation
The purpose of this improvement is to
automate several steps when a new release is created: the creation of regression labels on code and planning repos, the creation of bug report issue on planning repo, the addition of the bug report issue to the Releases Github Project board, and the setting of the RC cut date on the Github Project board. Finally this improvement ensures we no longer re-create the bug report issue when it already exists, which can happen when the release is re-cut.