test: rename `ganacheOptions` to generic `localNodeOptions` and removing option defaults from specs (#30058)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
In preparation to the [anvil migration
PR](https://github.com/MetaMask/metamask-extension/pull/27246), this PR
does the following:

- Renames `ganacheOptions` param in `withFixtures` to
`localNodeOptions`: the reason for that is because with a generic
option, we'll be able to use the same variable with whichever local node
we are using (being ganache, anvil, bitcoin, solana...in the future) --
see
[here](https://github.com/MetaMask/metamask-extension/pull/27246/files#r1939514183)
for more context
- Removes `ganacheOptions` from the specs where were using
`defaultGanacheOptions` and adds this to the ganache class : the reason
for that is because being those options already default, we don't need
to keep passing this variable in the specs, only in the case we want
custom ganache options

Note: the ganache class used one SRP, wherease the default ganache
options used a private key, which account doesn't derive from the same
SRP, leading to some specs using an account with balance 0, and some
others with balance 25ETH

We could either use the mnemonic or the account array in the options
object, but not both.
Our default fixtures use the imported priv key account, and given that
the majoritiy of the specs were using the default ganache options (the
imported priv key), the ganache class has been updated to use that, and
only in the few specs that were relying on the mnemonic, that has been
added as a custom option.
This also means that some balances are tweaked, as the fixtures use the
imported account in most of the specs, whereas if the ganache class is
used, that account will have 0 balance (now it will indeed have
balance).


Having this changes in a separate PR makes the other one smaller and
easier to review.



***Before***
We had 3 different cases:
- specs where they pass custom ganache options in ganacheOptions field,
and that override/extended the ganache class default
- specs where they passed default ganache options, which where defined
in a const in helpers, and that override/extended the ganache class
default
- specs where they didn't pass any option, or incorrectly passed
defaultGanache as a param in withFixtures, which then the ganache class
default options were applied

***Now***
We have consoldiated thej above in just:
- specs where they don't pass any ganacheOptions --> the default ganache
options will be applied, directly in the ganache class
- specs where they need custom options -> they will pass the
`localNodeOptions` flag and that override/extend the ganache class
default

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/30058?quickstart=1)

## **Related issues**

Fixes:

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: chloeYue <105063779+chloeYue@users.noreply.github.com>
Co-authored-by: Vinicius Stevam <45455812+vinistevam@users.noreply.github.com>
Co-authored-by: Salim TOUBAL <salim.toubal@outlook.com>
Co-authored-by: jake-perkins <128608287+jake-perkins@users.noreply.github.com>
Co-authored-by: Norbert Elter <72046715+itsyoboieltr@users.noreply.github.com>
Co-authored-by: Derek Brans <dbrans@gmail.com>
Co-authored-by: MetaMask Bot <metamaskbot@users.noreply.github.com>
Co-authored-by: Howard Braham <howrad@gmail.com>
Co-authored-by: cryptodev-2s <109512101+cryptodev-2s@users.noreply.github.com>
Co-authored-by: Matteo Scurati <matteo.scurati@gmail.com>
Co-authored-by: David Walsh <davidwalsh83@gmail.com>
Co-authored-by: David Drazic <david@timechaser.org>
Co-authored-by: António Regadas <apregadas@gmail.com>
Co-authored-by: Charly Chevalier <charly.chevalier@consensys.net>
Co-authored-by: Frederik Bolding <frederik.bolding@gmail.com>
Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>
Co-authored-by: Prithpal Sooriya <prithpal.sooriya@gmail.com>
Co-authored-by: Pedro Figueiredo <pedro.figueiredo@consensys.net>
Co-authored-by: Nidhi Kumari <nidhi.kumari@consensys.net>
Co-authored-by: sodofi <59624772+sodofi@users.noreply.github.com>
Co-authored-by: Christian Montoya <sirokai@gmail.com>
Co-authored-by: Mark Stacey <markjstacey@gmail.com>
Co-authored-by: Matthew Walsh <matthew.walsh@consensys.net>
Co-authored-by: Daniel <80175477+dan437@users.noreply.github.com>
Co-authored-by: Danica Shen <zhaodanica@gmail.com>
Co-authored-by: jiexi <jiexiluan@gmail.com>
Co-authored-by: Micaela Estabillo <100321200+micaelae@users.noreply.github.com>
Co-authored-by: sahar-fehri <sahar.fehri@consensys.net>
Co-authored-by: ffmcgee <51971598+ffmcgee725@users.noreply.github.com>
Co-authored-by: jpsains <32621022+jpsains@users.noreply.github.com>
Co-authored-by: Alex Donesky <adonesky@gmail.com>