chore: Expose E2E Chrome driver extension ID
The `extensionId` property was accidentally omitted by the ChromeDriver
builder. This was an old mistake, made in #7690. You can see this
property is expected in `test/e2e/webdriver/index.js`, though in
practice we don't use this `extensionId` for anything right now.

This was fixed so that we could use the `extensionId` in the future if
necessary. This was extracted from #27782, which used this ID in some
API spec e2e test helpers.