chore: Expose extension ID to E2E tests
The E2E test setup function (`withFixtures`) has been updated to pass
the extension ID to E2E tests. This will be useful in the near future
for testing the new multichain API, which is exposed over
`externally_connectable` and requires the extension ID to use.

This was extracted from #27782