test: fix `Send flow with some field validation` (#30808)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
This spec is flaky because we pass 2 mocks, the default solana mocks
(which include several mocks for apis and 50 SOL as balance) and another
custom one, specific for this test, where the balance is set to `0`. If
the test is fast enough, the `50` balance is not loaded and it passes,
but when the test is slower (FF), the `50` balance is loaded and then
the test fails.
The fix is to make that if the balance we pass is 0, this should
overwrite the balance 50 default mock, simply by changing the order of
the mocks.

![Screenshot from 2025-03-06
11-32-52](https://github.com/user-attachments/assets/24e0a2e8-8e08-4dbf-b38d-a9f41dc301c8)


![Screenshot from 2025-03-06
11-33-04](https://github.com/user-attachments/assets/16b197f2-514a-4db9-8b39-3b49ac018330)



[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/30808?quickstart=1)

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-extension/issues/30807

## **Manual testing steps**

1. Check ci

## **Screenshots/Recordings**


https://github.com/user-attachments/assets/91cbbb2c-f15e-499c-a862-494ecc95499b



## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.