Fix: Disable CSP override feature that is breaking sites
The CSP override feature has been temporarily disabled. The code that
activates the CSP override, as well as the CSP override state itself,
is still in place. But the activation code is no longer called, and the
settings page toggle for this feature has been removed.

We should follow up with a longer-term fix; either complete removal, or
fix the feature but default it to off (and migrate existing users to
start with this disabled).

Fixes #31094