test: add catch all mock and temporary allowlist (#26647)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
This PR does the following:
- adds a catch all mock, which will prevent any test to send real
requests to an API, and instead returns a 200
- displays logs anytime the request is going to a live server, or
redirected to the catch all mock--> for easy debugging
- adds a url and domain allowlist, which allows the requests to those
entries to be performed in the real servers. This list is temporary and
the goal is to reduce this to 0. However, adding this list was necessary
in order to make the specs that rely on those current live responses to
not fail and be able to merge this PR.

Merging this PR will make that any new request, is catched by the catch
all mock, so it already prevents new requests to happen.
The work on mocking all the whitelisted requests should be a joint
effort between teams.


[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/26647?quickstart=1)

## **Related issues**

Fixes: https://github.com/MetaMask/MetaMask-planning/issues/2762

## **Manual testing steps**

1. Check ci should work

## **Screenshots/Recordings**
When a request is sent to a live server, we'll see the following
information in the test logs:

![Screenshot from 2025-03-20
08-17-30](https://github.com/user-attachments/assets/a66d56a7-c4fe-4564-8f9f-3b60f80d2526)




I've enabled all test jobs just to be sure it doesn't break anything


https://app.circleci.com/pipelines/github/MetaMask/metamask-extension/132122/workflows/4ba5987a-28c3-41c3-8d5a-312dc88aacce

![Screenshot from 2025-03-19
16-57-18](https://github.com/user-attachments/assets/fc85f977-fbf6-4997-b9c6-79bd563fe919)

## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: Christian Montoya <sirokai@gmail.com>