fix: flaky test `Settings - general tab validate "मानक हिन्दी" language on tooltips` (#31384)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
In some of the change language specs, we change the language and
immediately click the close settings button. This action can have no
effect, if we click before the language is changes, as then there's a
re-render and we are still in the settings page.

We need to wait until the language has been changed before proceeding
with the next action. For that, we add a check to make sure the language
has changed, every time we perform a `changeLanguage` action.

![Screenshot from 2025-03-28
10-16-42](https://github.com/user-attachments/assets/bbded9ee-533e-4b34-bbdf-e7b05b3fa81f)


ci failure:
https://app.circleci.com/pipelines/github/MetaMask/metamask-extension/134139/workflows/c284f602-6903-4a81-a0ac-48a05f987ee7/jobs/4696447/tests

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/31384?quickstart=1)

## **Related issues**

Fixes:

## **Manual testing steps**

1. Check ci

## **Screenshots/Recordings**
See in the failure artifacts, how after clicking close settings, we
still remain in the settings page. The click was performed but took no
effect, as after that the page was re-rendered with the new language.
Then we try to look for an element from the home page but that can't be
found as we are still in settings



![image](https://github.com/user-attachments/assets/016d8cd0-13a4-4658-8979-e1d599e6a289)


## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.